Progreso de Pintado, Directorios y Limpieza de Codigo

+++ interfaz.py
@@ -1,5 +1,4 @@
 import os
-import os.path
 import platform
 import re
 import Lexico_JS
@@ -18,9 +17,7 @@ from CustomText import CustomText
 from TextLine import TextLineNumbers
 
 class Interfaz(tk.Frame):
-    ejecutar = None
-    
-
+    ejecutar = None    
     def __init__(self, *args, **kwargs):
         self.root = root
         tk.Frame.__init__(self, *args, **kwargs)
@@ -31,13 +28,10 @@ class Interfaz(tk.Frame):
         self.terminal = tk.Text(root, width=45, height=1, background="black",foreground="#00AA00")
         self.terminal.pack(side="right", fill="both", expand=True)
 
-
         # Special Text
         self.text = CustomText(self)
         self.vsb = tk.Scrollbar(orient="vertical", command=self.text.yview)
         self.text.configure(yscrollcommand=self.vsb.set)
-
-
 
         # Text line number
         self.linenumbers = TextLineNumbers(self, width=70)
@@ -57,7 +51,6 @@ class Interfaz(tk.Frame):
         report_dropdown = tk.Menu(menubar, tearoff=0)
         help_dropdown = tk.Menu(menubar, tearoff=0)
 
-
         file_dropdown.add_command(label="Nuevo", command=self.new_file)
         file_dropdown.add_command(label="Abrir", command=self.open_file)
         file_dropdown.add_command(label="Guardar", command=self.save)
@@ -69,19 +62,15 @@ class Interfaz(tk.Frame):
 
         report_dropdown.add_command(label="Errores Lexicos", command=self.errorReport)
         
-
         help_dropdown.add_command(label="Acerca de", command=self.about)
         help_dropdown.add_command(label="Manual de Usuario", command=self.m_user)
         help_dropdown.add_command(label="Manual Técnico", command=self.m_tecnic)
-
-
 
         menubar.add_cascade(label="Archivo", menu=file_dropdown)
         menubar.add_cascade(label="Ejecutar", menu=run_dropdown)
         menubar.add_cascade(label="Reportes", menu=report_dropdown)
         menubar.add_cascade(label="Ayuda", menu=help_dropdown)
-
-
+#-------------------------------------------------------Color Tags for the Paint Method---------------------------------------------------------------------
         self.text.tag_configure("reserved", foreground="red")
         self.text.tag_configure("var", foreground="#008000")
         self.text.tag_configure("int", foreground="#0000FF")
@@ -89,26 +78,16 @@ class Interfaz(tk.Frame):
         self.text.tag_configure("string", foreground="#FFFF00")
         self.text.tag_configure("comment", foreground="#808080")
         self.text.tag_configure("operator", foreground="#FFA500")
-
-
-
 #-------------------------------------------------------Line Number Method---------------------------------------------------------------------
     def _on_change(self, event):
         self.linenumbers.redraw()
         self.text.tag_remove('resaltado', '1.0', tk.END)
-        #lex = compiler.make_lexer()
-        #lex.input(self.text.get('1.0', tk.END))
-        
-        #self.pintar(txt)   
-
-
 #-------------------------------------------------------File Menu Methods---------------------------------------------------------------------
     def set_window_title(self, name=None):
         if name:
             self.root.title(name)
         else:
             self.root.title("Sin titulo.txt")
-
 
     def new_file(self):
         self.text.delete(1.0, tk.END)
@@ -123,7 +102,6 @@ class Interfaz(tk.Frame):
             with open(self.filename, "r") as f:
                self.text.insert(1.0, f.read())
             self.set_window_title(self.filename)
-
 
     def save(self):
         if self.filename:
@@ -152,11 +130,6 @@ class Interfaz(tk.Frame):
         value = messagebox.askokcancel("Salir", "Está seguro que desea salir?")
         if value :
                 root.destroy()
-            
-
-            
-
-
 #-------------------------------------------------------Execution Menu Methods---------------------------------------------------------------------       
     def analyze(self,entrada):
         js = Lexico_JS.lex_JS() 
@@ -170,23 +143,17 @@ class Interfaz(tk.Frame):
             self.terminal.insert(tk.END,str(js.token_output).replace("],", "\n").replace("[[","[").replace("]]","\n").replace("[","").replace("\\n","").replace("None,","").replace("None",""))
             self.terminal.insert(tk.END,"----------------------------------------Errors--------------------------------\n")
             self.terminal.insert(tk.END,str(js.error_output).replace("],", "\n").replace("[[","[").replace("]]","\n").replace("[",""))
-
+            self.error(js.error_list)
+            self.pintar(js.token_output)
+            if(str(txt).__contains__("PATHW:")):
+                self.get_direction("PATHW:",js.clean)
+                if(str(txt).__contains__("PATHL:")):
+                    self.get_direction("PATHL:",js.clean)
 
         else:
             box_tilte ="Execution Error"
             box_msg = "Por favor revise que el archivo que haya abierto sea JS, CSS, HTML o RMT"
-            messagebox.showerror(box_tilte,box_msg)
-        
-        self.error(js.error_list)
-        self.pintar(js.token_output)
-        if(str(txt).__contains__("PATHW:")): 
-            self.get_direction("PATHW:",js.clean)
-            if(str(txt).__contains__("PATHL:")):
-                self.get_direction("PATHL:",js.clean)
-        
-        
-                
-             
+            messagebox.showerror(box_tilte,box_msg)            
 #-------------------------------------------------------Help Menu Methods---------------------------------------------------------------------
     def about(self):
         box_tilte ="Autor"
@@ -195,30 +162,52 @@ class Interfaz(tk.Frame):
 
     def m_user(self):
         script_dir = os.path.dirname(os.path.abspath(__file__))
-        webbrowser.open_new(r'file://'+script_dir+'/WS.pdf')
+        direction = script_dir + "/WS.pdf" 
+        if (os.path.exists(direction)):
+            try:
+                webbrowser.open_new(r'file://'+direction)
+            except Exception as e:
+                box_tilte ="Path Error"
+                box_msg = "El archivo que trata de acceder no existe"
+                messagebox.showerror(box_tilte,box_msg)
+        else:
+            box_tilte ="New Path Error"
+            box_msg = "La ruta no existe"
+            messagebox.showerror(box_tilte,box_msg)
         
     def m_tecnic(self):
         script_dir = os.path.dirname(os.path.abspath(__file__))
-        webbrowser.open_new(r'file://'+script_dir+'/WS.pdf')
+        direction = script_dir + "/WS.pdf"
+        if (os.path.exists(direction)):
+            try:
+                webbrowser.open_new(r'file://'+direction)
+            except Exception as e:
+                box_tilte ="Path Error"
+                box_msg = "El archivo que trata de acceder no existe"
+                messagebox.showerror(box_tilte,box_msg)
+        else:
+            box_tilte ="New Path Error"
+            box_msg = "La ruta no existe"
+            messagebox.showerror(box_tilte,box_msg)
 #-------------------------------------------------------Descision Module Methods---------------------------------------------------------------------
     def verify_path(self):
         if self.filename == None:
-            box_tilte ="Path Error"
+            box_tilte ="File Extension Error"
             box_msg = "Por favor abra un archivo JS, CSS, HTML o RMT antes de analizar o escriba un archivo y guardelo con una de esas extensiones"
             messagebox.showerror(box_tilte,box_msg)
         else:
             self.path_module(self.filename)
   
     def path_module(self,entrada):
-        replaceS = entrada.replace("\\", "/")
-        x = replaceS.split(":",1)
-        if(x[0].isupper()):
-            var_split = os.path.splitext(replaceS)[1][1:]
+        if(os.path.exists(entrada)):
+            var_split = os.path.splitext(entrada)[1][1:]
+            print(var_split)
             self.decision_module(var_split)
         else:
-            var_split = os.path.splitext(replaceS)[1][1:]
-            self.decision_module(var_split)
-    
+            box_tilte ="Path File Error"
+            box_msg = "Archivo de entrada no existe"
+            messagebox.showerror(box_tilte,box_msg)
+               
     def decision_module(self,entrada):
         decision_flag = ""
         if(entrada == "html"):
@@ -237,40 +226,53 @@ class Interfaz(tk.Frame):
             decision_flag = "NONE"
             self.analyze(decision_flag)
 #-------------------------------------------------------File Management---------------------------------------------------------------------       
-    def create_file(self,path,clean):
-        fic = open(path+"file.js", "w")
-        
-        fic.write(clean)    
-        fic.close()
-
     def get_direction(self,entrada,clean):
         txt = self.text.get(1.0, tk.END)
         sistema = platform.system()
         if(entrada == "PATHW:" and sistema == "Windows"):            
-            path = txt.split(entrada)
-            direction = path[1].split("\n")
-            self.create_directory(direction[0],clean)
-        elif(entrada == "PATHL:" and sistema == "Linux"):
-            path = txt.split(entrada)
-            direction = path[1].split("\n")
-            self.create_directory(direction[0],clean)
+            if(txt.__contains__("PATHW: ")):
+                path = txt.split("PATHW: ")
+                direction = path[1].split("\n")
+                self.create_file(direction[0],clean)
+            elif(txt.__contains__("PATHW:")):
+                path = txt.split("PATHW:")
+                direction = path[1].split("\n")
+                self.create_file(direction[0],clean)               
+        elif(entrada == "PATHL:" and sistema == "Linux"):            
+            if(txt.__contains__("PATHL: ")):
+                path = txt.split("PATHL: ")
+                direction = path[1].split("\n")
+                self.create_file(direction[0],clean)
+            elif(txt.__contains__("PATHL:")):
+                path = txt.split("PATHL:")
+                direction = path[1].split("\n")
+                self.create_file(direction[0],clean) 
         else:
-            print("")
             box_tilte ="Operative System Error"
             box_msg = "La carpeta que esta tratando de crear no es el formato correcto en el sistema operativo en el que se encuentra actualmente"
             messagebox.showerror(box_tilte,box_msg)
 
- 
-    def create_directory(self,ruta,clean):
-        try:
-            os.makedirs(ruta)
-            self.create_file(ruta,clean)
-        except OSError:
-            pass
-  # si no podemos crear la ruta dejamos que pase
-  # si la operación resulto con éxito nos cambiamos al directorio 
-        os.chdir(ruta)
-        self.create_file(ruta,clean)
+    def create_file(self,path,clean):
+        if (os.path.exists(path)):
+            try:
+                os.chdir(path.replace("\\","/"))
+                fic = open(path+"file.js", "w")     
+                fic.write(clean)    
+                fic.close()
+            except Exception as e:
+                box_tilte ="Path Error"
+                box_msg = "La ruta del archivo que trata de acceder esta escrita de forma incorrecta, revise su archivo de entrada. Puede ser que tenga un espacio de más"
+                messagebox.showerror(box_tilte,box_msg)
+        else:
+            try:
+                os.makedirs(path.replace("\\","/"))
+                fic = open(path+"file.js", "w")     
+                fic.write(clean)    
+                fic.close()
+            except Exception as e:
+                box_tilte ="New Path Error"
+                box_msg = "La ruta del archivo que trata de acceder esta escrita de forma incorrecta, revise su archivo de entrada. Puede ser que tenga un espaciod de más"
+                messagebox.showerror(box_tilte,box_msg)
 #-------------------------------------------------------Reports---------------------------------------------------------------------       
     def error(self,entrada):
         if(len(entrada)==0):
@@ -283,48 +285,53 @@ class Interfaz(tk.Frame):
 
     def errorReport(self):
         script_dir = os.path.dirname(os.path.abspath(__file__))
-        webbrowser.open_new(r'file://' + script_dir + '/errorList.html')
-
+        direction = script_dir + "/errorList.html"
+        if(os.path.exists(direction)):
+            webbrowser.open_new(r'file://' + direction)
+        else:
+            box_tilte = "Report Error"
+            box_msg = "El archivo del reporte no existe"
+            messagebox.showinfo(box_tilte, box_msg)            
 #-------------------------------------------------------Paint Words---------------------------------------------------------------------       
     def pintar(self,token):
         for last in token:
-            if(last[2]=="reservada"):
-                posicionInicial = f'{last[0]}.{last[1]-1}'
-                posicionFinal = f'{posicionInicial}+{len(str(last[3]))}c'
-                self.text.tag_add('reserved', posicionInicial, posicionFinal)
+            if(last[0]!=None):
+                if(last[2]=="reservada"):
+                    posicionInicial = f'{last[0]}.{last[1]-1}'
+                    posicionFinal = f'{posicionInicial}+{len(str(last[3]))}c'
+                    self.text.tag_add('reserved', posicionInicial, posicionFinal)
 
-            elif(last[3].lower()=="var"):
-                posicionInicial = f'{last[0]}.{last[1]-1}'
-                posicionFinal = f'{posicionInicial}+{len(str(last[3]))}c'
-                self.text.tag_add('var', posicionInicial, posicionFinal)
+                elif(last[3].lower()=="var"):
+                    posicionInicial = f'{last[0]}.{last[1]-1}'
+                    posicionFinal = f'{posicionInicial}+{len(str(last[3]))}c'
+                    self.text.tag_add('var', posicionInicial, posicionFinal)
 
-            elif(last[2].lower()=="string"):
-                posicionInicial = f'{last[0]}.{last[1]-1}'
-                posicionFinal = f'{posicionInicial}+{len(str(last[3]))}c'
-                self.text.tag_add('string', posicionInicial, posicionFinal)
+                elif(last[2].lower()=="string"):
+                    posicionInicial = f'{last[0]}.{last[1]-1}'
+                    posicionFinal = f'{posicionInicial}+{len(str(last[3]))}c'
+                    self.text.tag_add('string', posicionInicial, posicionFinal)
 
-            elif(last[2].lower()=="integer"):
-                posicionInicial = f'{last[0]}.{last[1]-1}'
-                posicionFinal = f'{posicionInicial}+{len(str(last[3]))}c'
-                self.text.tag_add('int', posicionInicial, posicionFinal)
+                elif(last[2].lower()=="integer"):
+                    posicionInicial = f'{last[0]}.{last[1]-1}'
+                    posicionFinal = f'{posicionInicial}+{len(str(last[3]))}c'
+                    self.text.tag_add('int', posicionInicial, posicionFinal)
 
-            elif(last[3].lower()=="true" or last[3].lower()=="false"):
-                posicionInicial = f'{last[0]}.{last[1]-1}'
-                posicionFinal = f'{posicionInicial}+{len(str(last[3]))}c'
-                self.text.tag_add('boolean', posicionInicial, posicionFinal)
+                elif(last[3].lower()=="true" or last[3].lower()=="false"):
+                    posicionInicial = f'{last[0]}.{last[1]-1}'
+                    posicionFinal = f'{posicionInicial}+{len(str(last[3]))}c'
+                    self.text.tag_add('boolean', posicionInicial, posicionFinal)
 
-            elif(last[2].lower()=="comentario"):
-                posicionInicial = f'{last[0]}.{last[1]-1}'
-                posicionFinal = f'{posicionInicial}+{len(str(last[3]))}c'
-                self.text.tag_add('comment', posicionInicial, posicionFinal)
+                elif(last[2].lower()=="comentario"):
+                    posicionInicial = f'{last[0]}.{last[1]-1}'
+                    posicionFinal = f'{posicionInicial}+{len(str(last[3]))}c'
+                    self.text.tag_add('comment', posicionInicial, posicionFinal)
 
-            elif(last[2].lower()=="operador"):
-                posicionInicial = f'{last[0]}.{last[1]-1}'
-                posicionFinal = f'{posicionInicial}+{len(str(last[3]))}c'
-                self.text.tag_add('operator', posicionInicial, posicionFinal)
-
-
-
+                elif(last[2].lower()=="operador"):
+                    posicionInicial = f'{last[0]}.{last[1]-1}'
+                    posicionFinal = f'{posicionInicial}+{len(str(last[3]))}c'
+                    self.text.tag_add('operator', posicionInicial, posicionFinal)
+                else:
+                    pass
             else:
                 pass
 #-------------------------------------------------------Main---------------------------------------------------------------------
